adjustment on types, and moved products.json
